Visual C++ 2019 support

+++ internals/testing/dynamic_test.cpp
@@ -21,7 +21,7 @@ auto dynamic_test = test([]() {
 
   constexpr size_t n_values = 10;
   for (size_t i = 0; i < n_values; ++i)
-    atomically([&]() { queues[0].push_back(i); });
+    atomically([&]() { queues[0].push_back(static_cast<int>(i)); });
 
   for (size_t t = 0; t < n_threads; ++t) {
     std::thread([&]() {
